+++ breeze-core/src/test/kotlin/com/windea/breezeframework/core/extensions/StringExtensionsKtTest.kt
@@ -11,14 +11,14 @@ class StringExtensionsKtTest {
 		assertEquals(listOf("abc", "abc", "abc"), "__abc_abc__abc".splitToWordList('_'))
 		assertEquals(listOf("abc", "abc", "abc"), "--abc-abc--abc-".splitToWordList('-'))
 	}
-	
+
 	@Test //TESTED
 	fun toWordsTest() {
 		assertEquals("abc Abc", "abcAbc".toWords())
 		assertEquals("Abc Abc", "AbcAbc".toWords())
 		assertEquals("ABC Abc", "ABCAbc".toWords())
 	}
-	
+
 	@Test //TESTED
 	fun letterCaseTest() {
 		assertEquals("abcAbcAbc", "AbcAbcAbc".switchCaseBy(camelCase))
@@ -31,7 +31,7 @@ class StringExtensionsKtTest {
 		assertEquals("abcabc", "AbcAbc".switchCaseBy(lowercase))
 		assertEquals("ABCABC", "ABcABc".switchCaseBy(UPPERCASE))
 	}
-	
+
 	@Test //TESTED ALMOST
 	fun referenceCaseTest() {
 		assertEquals("Abc.Abc", "Abc.Abc".switchCaseBy(Standard))
@@ -39,34 +39,41 @@ class StringExtensionsKtTest {
 		assertEquals("$.abc.abc.[1].abc", "/abc/abc/1/abc".switchCaseBy(Json))
 		assertEquals("#/abc/abc/1/abc", "/abc/abc/1/abc".switchCaseBy(JsonSchema))
 	}
-	
+
 	@Test //TESTED
 	fun repeatExtensionTest() {
 		println("abc".repeat(3))
 		println("abc".flatRepeat(3))
 	}
-	
+
+	@Test
+	fun replaceExtensionTest() {
+		val string = "abc123abc123"
+		println(string.replace("""([a-z])\d""".toRegex(), "$1"))
+		println(string.replaceLooped("""([a-z])\d""".toRegex(), "$1"))
+	}
+
 	@Test //TESTED
 	fun messageFormatTest() {
-		assertEquals("123a123b123","123{0}123{1}123".messageFormat("a","b"))
+		assertEquals("123a123b123", "123{0}123{1}123".messageFormat("a", "b"))
 	}
-	
+
 	@Test //TESTED
 	fun customFormatTest() {
-		assertEquals("1a2b3","1{}2{}3".customFormat("{}","a","b"))
+		assertEquals("1a2b3", "1{}2{}3".customFormat("{}", "a", "b"))
 		assertEquals("1a2b3", "1\${}2\${}3".customFormat("\${}", "a", "b"))
 		assertEquals("1a2b3","1{0}2{1}3".customFormat("{index}","a","b"))
 		assertEquals("1a2b3", "1{aaa}2{bbb}3".customFormat("{name}", "aaa" to "a", "bbb" to "b"))
 		assertEquals("1b2a3","1{1}2{0}3".customFormat("{index}","a","b"))
 		assertEquals("1b2a3", "1{bbb}2{aaa}3".customFormat("{name}", "aaa" to "a", "bbb" to "b"))
 	}
-	
+
 	@Test //TESTED
 	fun quoteTest() {
 		assertEquals(""""'1\"2'"""", """'1"2'""".wrapQuote('"'))
 		assertEquals("""'1"2'""", """"'1\"2'"""".unwrapQuote())
 	}
-	
+
 	@Test //TESTED
 	fun progressiveTest() {
 		val a = """
@@ -78,7 +85,7 @@ class StringExtensionsKtTest {
 			123123
 			123123123
 		""".trimIndent()
-		
+
 		//println(a plusByLine b)
 		//println(a.padStartByLine())
 		//println(a.padEndByLine())