+++ CHANGELOG.md
@@ -237,6 +237,7 @@
 * [X] breeze-core 添加`String.orEmpty(predicate)`、`String.takeIfNotEmpty`和`String.takeIfNotBlank`。
 * [X] breeze-core&breeze-reflect 优化部分代码。
 * [X] 添加必要的`@JvmStatic`和`@JvmOverloads`注解。
+* [X] 添加`String.replaceLooped`扩展。
 
 # 待办